requirements.txt
Django==3.1.7
djangorestframework
pillow==8.1.2